Use `PositionFor` instead of `Position` (#7)
If a node has a `//line` comment the position reported from `Position`
will be based on that comment. However, when we suggest fixes in
`golanngci-lint` we need to do this in the current file, even if it's
generated.

This will ensure we get the actual position in the potentially generated
file and ignoring the comment directive.