Make linter work on non formatted files
Instead of relying on the file being `gofmt`:ed and only containing one
empty line at most this commit will keep track of comments after the
opening bracket of block statements and set the fix start to where this
comment ends. This means that we will fix all the way from the left
bracket or comment til the first statement no matter how many empty
lines this is.

This will also keep a list of lines affected by the linter to support
fixing multiple empty lines in `golangci-lint`.